requirements.txt
discord.py>=2.3.0
python-dotenv>=1.0.0
resvg-py>=0.3.0
faster-whisper>=1.0.0
numpy>=1.24.0
pypinyin>=0.50.0
pytest>=7.0.0
